refactor: fetchData 함수 index.ts -> util.ts 이동했습니다!
