+++ apps/client/services/index.ts
@@ -4,45 +4,6 @@ export * from './kakao';
 export * from './like';
 export * from './mbti';
 
-type Method = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
-
-interface FetchOptions {
-  headers: Headers;
-  body: string;
-  params: object;
-  method: Method;
-}
-
-interface FetchOptionProps extends Omit<FetchOptions, 'body' | 'method'> {
-  body: { [key: string]: any };
-}
-
-export const fetchData = async <T>(url: string, method: Method, options?: Partial<FetchOptionProps>) => {
-  const { body, ...restOptions } = options ?? {};
-  const fetchOptions = {
-    method,
-    ...restOptions,
-    ...(body ? { body: JSON.stringify(body) } : undefined),
-  };
-
-  const response = await fetch(`${process.env.NEXT_PUBLIC_BE_URL_PROD}${url}`, fetchOptions);
-
-  if (!response.ok) {
-    switch (response.status) {
-      case 404:
-        return notFound();
-      case 401:
-        return;
-      default:
-        throw new Error('Failed to fetch data');
-    }
-  }
-
-  if (response.status === 204) return;
-
-  return response.json() as T;
-};
-
 // 기존
 export const fetchClient = async ({ url, method, headers, body }: Services.FetchClientProp) => {
   const isInvaildUrl = !url || typeof url !== 'string';
