+++ apps/client/services/util.ts
@@ -1,3 +1,5 @@
+import { notFound } from 'next/navigation';
+
 import { LOGIN } from '@/constants/constant';
 
 interface CreateHeadersProrps {
@@ -5,7 +7,19 @@ interface CreateHeadersProrps {
   cacheControl?: string;
 }
 
-// creatHeaders 추가
+type Method = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH';
+
+interface FetchOptions {
+  headers: Headers;
+  body: string;
+  params: object;
+  method: Method;
+}
+
+interface FetchOptionProps extends Omit<FetchOptions, 'body' | 'method'> {
+  body: { [key: string]: any };
+}
+
 export const createHeaders = (headerAppend?: CreateHeadersProrps) => {
   const { contnetType, cacheControl } = headerAppend ?? {};
   const headers = new Headers();
@@ -20,4 +34,30 @@ export const createHeaders = (headerAppend?: CreateHeadersProrps) => {
   headers.append('Authorization', token);
 
   return headers;
+};
+
+export const fetchData = async <T>(url: string, method: Method, options?: Partial<FetchOptionProps>) => {
+  const { body, ...restOptions } = options ?? {};
+  const fetchOptions = {
+    method,
+    ...restOptions,
+    ...(body ? { body: JSON.stringify(body) } : undefined),
+  };
+
+  const response = await fetch(`${process.env.NEXT_PUBLIC_BE_URL_PROD}${url}`, fetchOptions);
+
+  if (!response.ok) {
+    switch (response.status) {
+      case 404:
+        return notFound();
+      case 401:
+        return;
+      default:
+        throw new Error('Failed to fetch data');
+    }
+  }
+
+  if (response.status === 204) return;
+
+  return response.json() as T;
 };
